refactor: move event on level above
